Upload excel

+++ models/stock_model.php
@@ -1,4 +1,5 @@
 <?php
+include_once('libs/PHPExcel.php');
 
 class Stock_Model extends Model {
 
@@ -79,6 +80,71 @@ class Stock_Model extends Model {
 
                 $this->db->commit();
                 fclose($file);
+                unlink($filename);
+            } else {
+                $result = "0";
+                $error = "No upload file";
+            }
+
+        } catch (Exception $e) {
+            $result = "0";
+            $error = $e->getMessage();
+            $this->db->rollBack();
+        }
+        
+        $data = array('res' => $result, 'error' => $error);
+        echo json_encode($data);
+
+    }
+
+    public function uploadStockSystemExcel() {
+
+        $result = "1";
+        $error = "";
+
+        try {
+            $this->db->beginTransaction();
+
+            if (isset($_FILES['stkfile'])) {
+
+                $filename = $_FILES['stkfile']['name'];
+                
+                if($_FILES["stkfile"]["size"] > 0)
+                {
+                    move_uploaded_file($_FILES["stkfile"]["tmp_name"],$filename);
+                    $excelReader = PHPExcel_IOFactory::createReaderForFile($filename);
+                    $excelObj = $excelReader->load($filename);
+                    $worksheet = $excelObj->getSheet(0);
+                    $lastRow = $worksheet->getHighestRow();
+
+                    $sql = "DELETE FROM stock_import";
+                    $stmt = $this->db->prepare($sql);
+                    $stmt->execute();
+
+                    $sql = "INSERT INTO stock_import values (:id, :name, :qty)";
+                    $stmt = $this->db->prepare($sql);
+                    // echo $file;
+                    for ($row = 13; $row <= $lastRow; $row++) {
+                        $stmt->execute(array(
+                            ':id'=>$worksheet->getCell('B'.$row)->getValue()
+                            ,':name'=>$worksheet->getCell('C'.$row)->getValue()
+                            ,':qty'=>$worksheet->getCell('D'.$row)->getValue()
+                        ));
+                    }
+
+                    $sql = "REPLACE INTO stock (stkdate, stkitmgrp, stkitmcd, stkoutqty, stkroomqty, stksysqty, stkunt, stkadjqty)
+                            SELECT CURRENT_DATE, p.pmdval1, p.pmdcd, stkoutqty, stkroomqty, stiqty, p.pmdval4, stkadjqty
+                            FROM (SELECT * FROM prmdtl WHERE pmdtbno=12) p
+                            LEFT JOIN stock_import ON REPLACE(stiname,' (ราคาปกติ)','')=p.pmddesc
+                            LEFT JOIN stock ON stkdate=CURRENT_DATE 
+                                    AND p.pmdcd=stkitmcd";
+                    $stmt = $this->db->prepare($sql);
+                    $stmt->execute();
+                }
+
+                $error = $filename;
+
+                $this->db->commit();
                 unlink($filename);
             } else {
                 $result = "0";
